requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
prompt_toolkit==3.0.52
Pygments==2.19.2
python-dotenv==1.1.1
python-magic==0.4.27
questionary==2.1.1
rich==14.1.0
tabulate==0.9.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
